Exclude profiles pre-1900
As mentioned #165 CoTeDe uses a datetime function that cannot handle dates before 1900. This filters out those observations from being put in the database.

Alternative is to leave them in and exclude them from going through the quality control in AutoQC.py. However that could cause confusion later when analysing results and it is straightforward to add extra rows to the database at a later time if we want to process them.